# backend/docs/openapi.yaml
openapi: 3.1.0
info:
  title: AVRA VIVA Marketplace API
  version: 0.2.0
  description: Backend REST contract. Import workflows are administrative CLI operations and are documented separately.
servers:
  - url: /
...
paths:
...
  /api/v1/systems/{id}:
    get:
      tags: [Catalogue]
      summary: Get system
      parameters: [ { $ref: '#/components/parameters/Id' } ]
      responses:
        '200': { $ref: '#/components/responses/SystemItem' }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/v1/systems/by-code/{code}:
    get:
      tags: [Catalogue]
      summary: Get system by functional code
      parameters: [ { $ref: '#/components/parameters/Code' } ]
      responses:
        '200': { $ref: '#/components/responses/SystemItem' }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/v1/subsystems:
    get:
      tags: [Catalogue]
      summary: List subsystems
      parameters:
        - name: system_id
          in: query
          schema: { type: string, format: uuid }
      responses:
        '200': { description: Success }
  /api/v1/subsystems/{id}:
    get:
      tags: [Catalogue]
      summary: Get subsystem
      parameters: [ { $ref: '#/components/parameters/Id' } ]
      responses:
        '200': { description: Success }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/v1/subsystems/by-code/{code}:
    get:
      tags: [Catalogue]
      summary: Get subsystem by functional code
      parameters: [ { $ref: '#/components/parameters/Code' } ]
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { $ref: '#/components/schemas/Subsystem' }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/v1/archetypes:
    get:
      tags: [Catalogue]
      summary: List archetypes
      parameters:
        - name: subsystem_id
          in: query
          schema: { type: string, format: uuid }
      responses:
        '200': { description: Success }
  /api/v1/archetypes/{id}:
    get:
      tags: [Catalogue]
      summary: Get archetype
      parameters: [ { $ref: '#/components/parameters/Id' } ]
      responses:
        '200': { description: Success }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/v1/archetypes/by-code/{code}:
    get:
      tags: [Catalogue]
      summary: Get archetype by functional code
      parameters: [ { $ref: '#/components/parameters/Code' } ]
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { $ref: '#/components/schemas/Archetype' }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/v1/generic-solutions:
    get:
      tags: [Catalogue]
      summary: List generic solutions
      parameters:
        - name: archetype_id
          in: query
          schema: { type: string, format: uuid }
      responses:
        '200': { description: Success }
  /api/v1/generic-solutions/by-code/{code}:
    get:
      tags: [Catalogue]
      summary: Get a generic solution by functional code with Slots and Assets
      parameters: [ { $ref: '#/components/parameters/Code' } ]
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { $ref: '#/components/schemas/GenericSolution' }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/v1/generic-solutions/{id}:
    get:
      tags: [Catalogue]
      summary: Get a generic solution with Slots and Assets
      parameters: [ { $ref: '#/components/parameters/Id' } ]
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { $ref: '#/components/schemas/GenericSolution' }
        '404': { $ref: '#/components/responses/NotFound' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }
    Code:
      name: code
      in: path
      required: true
      schema: { type: string }
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: { type: string }
            message: { type: string }
    System:
      type: object
      required: [id, code, name_es, created_at, updated_at]
      properties:
        id: { type: string, format: uuid }
        code: { type: string }
        name_es: { type: string }
        description: { type: [string, 'null'] }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    Subsystem:
      type: object
      required: [id, system_id, code, name_es, created_at, updated_at]
      properties:
        id: { type: string, format: uuid }
        system_id: { type: string, format: uuid }
        code: { type: string }
        name_es: { type: string }
        description: { type: [string, 'null'] }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    Archetype:
      type: object
      required: [id, subsystem_id, code, name_es, created_at, updated_at]
      properties:
        id: { type: string, format: uuid }
        subsystem_id: { type: string, format: uuid }
        code: { type: string }
        name_es: { type: string }
        description: { type: [string, 'null'] }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    Asset:
      type: object
      properties:
        id: { type: string, format: uuid }
        code: { type: [string, 'null'] }
        original_filename: { type: string }
        mime_type: { type: string }
        size: { type: integer }
        sha256: { type: string }
        asset_type: { type: string }
        format: { type: string }
        role: { type: string }
        validation_data: { type: [object, 'null'], additionalProperties: true }
        extraction_data: { type: [object, 'null'], additionalProperties: true }
    Slot:
      type: object
      properties:
        id: { type: string, format: uuid }
        key: { type: string }
        name_es: { type: string }
        role: { type: [string, 'null'] }
        sequence: { type: integer }
        required: { type: boolean }
        properties: { type: object, additionalProperties: true }
        metrics: { type: [object, 'null'], additionalProperties: true }
        source_reference: { type: [object, 'null'], additionalProperties: true }
    GenericSolution:
      type: object
      properties:
        id: { type: string, format: uuid }
        archetype_id: { type: string, format: uuid }
        code: { type: string }
        name_es: { type: string }
        description: { type: [string, 'null'] }
        functional_unit: { type: [string, 'null'] }
        status: { type: [string, 'null'] }
        classifications: { type: array, items: { type: object, additionalProperties: true } }
        source_references: { type: array, items: { type: object, additionalProperties: true } }
        attributes: { type: object, additionalProperties: true }
        metrics: { type: object, additionalProperties: true }
        cte_compliance: { type: object, additionalProperties: true }
        environmental_data: { type: object, additionalProperties: true }
        economic_data: { type: object, additionalProperties: true }
        industrialization_data: { type: object, additionalProperties: true }
        viva_metrics: { type: object, additionalProperties: true }
        slots: { type: array, items: { $ref: '#/components/schemas/Slot' } }
        assets: { type: array, items: { $ref: '#/components/schemas/Asset' } }
...
  responses:
...
    SystemItem:
      description: Success
      content:
        application/json:
          schema:
            type: object
            properties:
              data: { $ref: '#/components/schemas/System' }
    NotFound:
      description: Not found
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...